+++ checkout-tabs-wp-ml.php
@@ -3,7 +3,7 @@
  * Plugin Name: Checkout Tabs WP ML
  * Plugin URI: https://github.com/caslusilver/checkout-tabs-wp-ml-dev
  * Description: Abas no checkout do WooCommerce (checkout clássico) com consulta de frete via webhook e integração em WC_Session.
- * Version: 3.1.34
+ * Version: 3.2.0
  * Author: Lucas Andrade / AI
  * Author URI: https://github.com/caslusilver
  * License: GPL2
@@ -29,6 +29,7 @@ require_once CHECKOUT_TABS_WP_ML_DIR . 'inc/frontend/addresses-api.php';
 require_once CHECKOUT_TABS_WP_ML_DIR . 'inc/enqueue.php';
 require_once CHECKOUT_TABS_WP_ML_DIR . 'inc/ajax-signup.php';
 require_once CHECKOUT_TABS_WP_ML_DIR . 'inc/cpf-lock.php';
+require_once CHECKOUT_TABS_WP_ML_DIR . 'inc/ajax-debug-logs.php';
 require_once CHECKOUT_TABS_WP_ML_DIR . 'inc/ajax-store-webhook-shipping.php';
 require_once CHECKOUT_TABS_WP_ML_DIR . 'inc/shipping-rates-override.php';
 require_once CHECKOUT_TABS_WP_ML_DIR . 'inc/admin-refresh-cache.php';